python/requirements.txt
fastapi
uvicorn
python-multipart
scenedetect
opencv-python-headless
moviepy
openai-whisper
sentence-transformers
Pillow
numpy